// src/sharedmodule/interface/Interface.tsx
export interface QuizQuestionPreview {
  _id: string;
  title: string;
  options: QuestionOptions;
}
// ...
export interface UpcommingQuizzProps {
  _id: string;
  code: string;
  title: string;
  description: string;
  status: "active" | "closed"; // or string if open-ended
  instructor: string;
  group: string;
  questions_number: number;
  questions: QuizQuestionPreview[];
  schadule: string; // ISO date string
  duration: number; // in minutes or hours
  score_per_question: number;
  type: string; // e.g. "FE", "BE", etc.
  difficulty: "easy" | "medium" | "hard";
  updatedAt: string;
  createdAt: string;
  __v: number;
  closed_at: string;
  participants: number;
}
// ...
export interface QuizzGroupList{
    _id: string;
    name: string;
}
// ...
export interface CreateAndUpdateQuizzProps {
  title: string;
  description: string;
  group: string; // Group ID
  questions_number: number;
  difficulty: "easy" | "medium" | "hard";
  type: "FE" | "BE"|"DB";
  schadule: string; // ISO string: "2024-02-15T21:19:34"
  duration: number; // in minutes
  score_per_question: number;
}
export interface QuizzAddAndUpdateProps{
  title:string | null;
  errorMessage:string| null
  onCreate:(dataInfo:CreateAndUpdateQuizzProps)=>Promise<void>;
  onUpdate:(dataInfo:CreateAndUpdateQuizzProps,quizzId:string)=>Promise<void>
  onClose:()=>void;
  currentQuizz:UpcommingQuizzProps | null;
}